+++ tests/test_request_2_0.py
@@ -1,12 +1,13 @@
 """Test encoding and decoding of iiif request URLs
 
-Follows http://www-sul.stanford.edu/iiif/image-api/ v1.0 dated "10 August 2012"
+Follows http://iiif.io/api/image/2.0/
 
 This test includes a number of test cases beyond those given as examples
 in the table in section 7 of the spec. See iiif_urltest_spec.py for the
 set given in the spec.
 
 Simeon Warner - 2012-03-23, 2012-04-13
+  2014-11-01 - Updated for v2.0 spec
 """
 import unittest
 
@@ -18,66 +19,63 @@ from iiif.request import IIIFRequest,IIIFRequestBaseURI
 # 
 data  = {
     '00_params': [
-        {'identifier':'id1', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'native' },
-        'id1/full/full/0/native'],
+        {'identifier':'id1', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'default' },
+        'id1/full/full/0/default'],
     '02_params': [
-        {'identifier':'id1', 'region':'full', 'size':'100,100', 'rotation':'0', 'quality':'native', 'format':'jpg' },
-        'id1/full/100,100/0/native.jpg'],
+        {'identifier':'id1', 'region':'full', 'size':'100,100', 'rotation':'0', 'quality':'default', 'format':'jpg' },
+        'id1/full/100,100/0/default.jpg'],
     '03_params': [
-        {'identifier':'id1', 'region':'full', 'size':'100,100', 'rotation':'0', 'quality':'grey' },
-        'id1/full/100,100/0/grey'],
+        {'identifier':'id1', 'region':'full', 'size':'100,100', 'rotation':'0', 'quality':'gray' },
+        'id1/full/100,100/0/gray'],
     '04_params': [
-        {'identifier':'id1', 'region':'full', 'size':'100,100', 'rotation':'0', 'quality':'grey', 'format':'tif' },
-        'id1/full/100,100/0/grey.tif'],
+        {'identifier':'id1', 'region':'full', 'size':'100,100', 'rotation':'0', 'quality':'gray', 'format':'tif' },
+        'id1/full/100,100/0/gray.tif'],
     '05_params': [
         {'identifier':'bb157hs6068', 'region':'full', 'size':'pct:100', 'rotation':'270', 'quality':'bitonal', 'format':'jpg'},
         'bb157hs6068/full/pct:100/270/bitonal.jpg',
         'bb157hs6068/full/pct%3A100/270/bitonal.jpg'],
     '06_params': [
-        {'identifier':'bb157hs6068', 'region':'full', 'size':'100,', 'rotation':'123.456', 'quality':'grey', 'format':'jpg'},
-        'bb157hs6068/full/100,/123.456/grey.jpg'],
+        {'identifier':'bb157hs6068', 'region':'full', 'size':'100,', 'rotation':'123.456', 'quality':'gray', 'format':'jpg'},
+        'bb157hs6068/full/100,/123.456/gray.jpg'],
 # ARKs from http://tools.ietf.org/html/draft-kunze-ark-00
 # ark:sneezy.dopey.com/12025/654xz321
 # ark:/12025/654xz321
     '21_ARK ': [
-        {'identifier':'ark:sneezy.dopey.com/12025/654xz321', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'native'},
-        'ark:sneezy.dopey.com%2F12025%2F654xz321/full/full/0/native',
-        'ark%3Asneezy.dopey.com%2F12025%2F654xz321/full/full/0/native'],
+        {'identifier':'ark:sneezy.dopey.com/12025/654xz321', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'default'},
+        'ark:sneezy.dopey.com%2F12025%2F654xz321/full/full/0/default',
+        'ark%3Asneezy.dopey.com%2F12025%2F654xz321/full/full/0/default'],
     '22_ARK ': [
-        {'identifier':'ark:/12025/654xz321', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'native'},
-        'ark:%2F12025%2F654xz321/full/full/0/native',
-        'ark%3A%2F12025%2F654xz321/full/full/0/native'],
+        {'identifier':'ark:/12025/654xz321', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'default'},
+        'ark:%2F12025%2F654xz321/full/full/0/default',
+        'ark%3A%2F12025%2F654xz321/full/full/0/default'],
 # URNs from http://tools.ietf.org/html/rfc2141
 # urn:foo:a123,456
     '31_URN ': [
-        {'identifier':'urn:foo:a123,456', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'native'},
-        'urn:foo:a123,456/full/full/0/native',
-        'urn%3Afoo%3Aa123,456/full/full/0/native'],
+        {'identifier':'urn:foo:a123,456', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'default'},
+        'urn:foo:a123,456/full/full/0/default',
+        'urn%3Afoo%3Aa123,456/full/full/0/default'],
 # URNs from http://en.wikipedia.org/wiki/Uniform_resource_name
 # urn:sici:1046-8188(199501)13:1%3C69:FTTHBI%3E2.0.TX;2-4
 # ** note will get double encoding **
     '32_URN ': [
-        {'identifier':'urn:sici:1046-8188(199501)13:1%3C69:FTTHBI%3E2.0.TX;2-4', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'native'},
-        'urn:sici:1046-8188(199501)13:1%253C69:FTTHBI%253E2.0.TX;2-4/full/full/0/native',
-        'urn%3Asici%3A1046-8188(199501)13%3A1%253C69%3AFTTHBI%253E2.0.TX;2-4/full/full/0/native'],
+        {'identifier':'urn:sici:1046-8188(199501)13:1%3C69:FTTHBI%3E2.0.TX;2-4', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'default'},
+        'urn:sici:1046-8188(199501)13:1%253C69:FTTHBI%253E2.0.TX;2-4/full/full/0/default',
+        'urn%3Asici%3A1046-8188(199501)13%3A1%253C69%3AFTTHBI%253E2.0.TX;2-4/full/full/0/default'],
 # Extreme silliness
     '41_odd ': [
-        {'identifier':'http://example.com/?54#a', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'native'},
-        'http:%2F%2Fexample.com%2F%3F54%23a/full/full/0/native',
-        'http%3A%2F%2Fexample.com%2F?54#a/full/full/0/native'],
+        {'identifier':'http://example.com/?54#a', 'region':'full', 'size':'full', 'rotation':'0', 'quality':'default'},
+        'http:%2F%2Fexample.com%2F%3F54%23a/full/full/0/default',
+        'http%3A%2F%2Fexample.com%2F?54#a/full/full/0/default'],
     # Info requests
     '50_info': [
         {'identifier':'id1', 'info':True, 'format':'json' },
         'id1/info.json'],
-    '51_info': [
-        {'identifier':'id1', 'info':True, 'format':'xml' },
-        'id1/info.xml'],
     }
 
 class TestAll(unittest.TestCase):
 
     def test01_parse_region(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.region=None
         r.parse_region()
         self.assertTrue(r.region_full)
@@ -106,7 +104,7 @@ class TestAll(unittest.TestCase):
         self.assertEqual(r.region_xywh, [0.0,0.0,100.0,100.0])
 
     def test02_parse_region_bad(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.region='pct:0,0,50,1000'
         self.assertRaises( IIIFError, r.parse_region )
         r.region='pct:-10,0,50,100'
@@ -115,7 +113,7 @@ class TestAll(unittest.TestCase):
         self.assertRaises( IIIFError, r.parse_region )
 
     def test03_parse_size(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.parse_size('pct:100')
         self.assertEqual(r.size_pct, 100.0)
         self.assertFalse(r.size_bang)
@@ -137,14 +135,14 @@ class TestAll(unittest.TestCase):
         self.assertEqual(r.size_wh, (5,6))
 
     def test04_parse_size_bad(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         self.assertRaises( IIIFError, r.parse_size, ',0.0' )
         self.assertRaises( IIIFError, r.parse_size, '0.0,' )
         self.assertRaises( IIIFError, r.parse_size, '1.0,1.0' )
         self.assertRaises( IIIFError, r.parse_size, '1,1,1' )
 
     def test05_parse_rotation(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.parse_rotation('0')
         self.assertEqual(r.rotation_mirror, False)
         self.assertEqual(r.rotation_deg, 0.0)
@@ -171,7 +169,7 @@ class TestAll(unittest.TestCase):
         self.assertEqual(r.rotation_deg, 123.45678)
 
     def test06_parse_rotation_bad(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.rotation='-1'
         self.assertRaises( IIIFError, r.parse_rotation )
         r.rotation='-0.0000001'
@@ -186,22 +184,22 @@ class TestAll(unittest.TestCase):
         self.assertRaises( IIIFError, r.parse_rotation )
 
     def test07_parse_quality(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.quality=None
         r.parse_quality()
-        self.assertEqual(r.quality_val, 'native')
-        r.quality='native'
+        self.assertEqual(r.quality_val, 'default')
+        r.quality='default'
         r.parse_quality()
-        self.assertEqual(r.quality_val, 'native')
+        self.assertEqual(r.quality_val, 'default')
         r.quality='bitonal'
         r.parse_quality()
         self.assertEqual(r.quality_val, 'bitonal')
-        r.quality='grey'
+        r.quality='gray'
         r.parse_quality()
-        self.assertEqual(r.quality_val, 'grey')
+        self.assertEqual(r.quality_val, 'gray')
 
     def test08_parse_quality_bad(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         r.quality='does_not_exist'
         self.assertRaises( IIIFError, r.parse_quality )
         # bad ones
@@ -212,7 +210,7 @@ class TestAll(unittest.TestCase):
         for tname in sorted(data.iterkeys()):
             tdata=data[tname]
             print tname + "   " + self.pstr(data[tname][0]) + "  " + data[tname][1]
-            iiif = IIIFRequest(api_version='1.0',**data[tname][0])
+            iiif = IIIFRequest(**data[tname][0])
             self.assertEqual(iiif.url(),data[tname][1])
         print
   
@@ -221,7 +219,7 @@ class TestAll(unittest.TestCase):
             tdata=data[tname]
             pstr = self.pstr(data[tname][0])
             for turl in data[tname][1:]:
-                iiif = IIIFRequest(api_version='1.0').split_url(turl)
+                iiif = IIIFRequest().split_url(turl)
                 tstr = self.pstr(iiif.__dict__)
                 print tname + "   " + turl + " -> " + tstr
                 self.assertEqual(tstr,pstr)
@@ -235,16 +233,16 @@ class TestAll(unittest.TestCase):
         self.assertRaises(IIIFError, IIIFRequest().split_url, ("id1/all/270/!pct%3A75.23.jpg"))
 
     def test20_parse_w_comma_h(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         self.assertEquals( r._parse_w_comma_h('1,2','a'), (1,2) )
 
     def test21_parse_w_comma_h_bad(self):
-        r = IIIFRequest(api_version='1.0')
+        r = IIIFRequest()
         self.assertRaises( IIIFError, r._parse_w_comma_h, '1.0,1.0', 'size' )
 
     def pstr(self,p):
         s=''
-        for k in ['identifier','region','size','rotation','native','info','format']:
+        for k in ['identifier','region','size','rotation','default','info','format']:
             if k in p and p[k]:
                 s += k+'='+str(p[k])+' '
         return(s)